Se integran los 2 componentes de vista prestación en uno
