--- src/app/modules/rup/components/ejecucion/vistaPrestacion.html
@@ -1,73 +1,0 @@
-<ng-container *ngIf="registro">
-    <h4 class=" text-capitalize">{{
-        registro.solicitud.tipoPrestacion.term }}</h4>
-    <div class="row">
-        <div class="col-6">
-            <b> Fecha: </b>
-            {{ registro.ejecucion.fecha | date: 'EEE dd/MM/yyyy HH:MM' }}
-        </div>
-        <div class="col-6" *ngIf="(registro.ejecucion.fecha - registro.solicitud.fecha) > 1000 ">
-            <b> Solicitada el: </b>
-            {{ registro.solicitud.fecha | date: 'EEE dd/MM/yyyy HH:MM' }}
-        </div>
-        <div class="col-12">
-            <b> Solicitada por el profesional: </b>
-            {{ registro.solicitud.profesional | profesional }}
-        </div>
-        <div class="col-12">
-            <b> Desde la Organización: </b>
-            {{ registro.solicitud.organizacion.nombre }}
-        </div>
-    </div>
-    <div class="row">
-        <div class="col-12">
-            <hr>
-            <h4>Registros de la consulta:</h4>
-            <ng-container *ngFor="let elemento of registro.ejecucion.registros; let i = index">
-                <div class="rup-card {{ elemento.esSolicitud ? 'solicitud' : servicioPrestacion.getCssClass(elemento.concepto) }}">
-                    <div class="rup-header">
-                        <div class="icon-rup">
-                            <i class="icon icon-rup-semantic-{{servicioPrestacion.getIcon(elemento.concepto)}}"></i>
-                        </div>
-                        <div class="title">
-                            <ng-container *ngIf="elemento?.nombre">
-                                {{elemento.nombre[0].toUpperCase() +
-                                elemento.nombre.slice(1)}}
-                            </ng-container>
-                        </div>
-                        <div class="actions">
-                            <small *ngIf="elemento?.relacionadoCon && elemento?.relacionadoCon.length > 0" class="badge badge-info">
-                                <b>Relacionado con:</b> {{
-                                elemento.relacionadoCon[0].concepto.term[0].toUpperCase()
-                                + elemento.relacionadoCon[0].concepto.term.slice(1)}}
-                            </small>
-                            <small class="badge badge-info" *ngIf="elemento.esDiagnosticoPrincipal">
-                                Procedimiento / diagnóstico principal
-                            </small>
-                        </div>
-                    </div>
-                    <div class="rup-body">
-                        <div class="legend">
-                            <span>
-                                {{ (elemento.esSolicitud) ? 'solicitud' :
-                                elemento.concepto.semanticTag}}
-                            </span>
-                        </div>
-                        <div class="content">
-                            <div class="row">
-
-                                <div class="col-12" *ngIf="paciente">
-                                    <rup [elementoRUP]="elementosRUPService.buscarElemento(elemento.concepto, elemento.esSolicitud)"
-                                        [params]="elementosRUPService.selectPorRefsetId(elemento.concepto)"
-                                        [prestacion]="registro" [registro]="elemento" [paciente]="paciente"
-                                        [soloValores]="true"></rup>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </ng-container>
-        </div>
-    </div>
-
-</ng-container>
