+++ src/app/modules/rup/components/huds/vistaPrestacion.html
@@ -1,4 +1,4 @@
-<div class="prestaciones">
+<div class="prestaciones mt-3 ml-3">
     <ng-container *ngIf="prestacion">
         <h4 class="text-capitalize">{{ prestacion.solicitud.tipoPrestacion.term }}</h4>
         <div class="row">
@@ -27,10 +27,11 @@
                 <ng-container *ngFor="let elemento of prestacion.ejecucion.registros">
 
                     <div
-                        class="rup-card {{ elemento.esSolicitud ? 'solicitud' : prestacionesService.getCssClass(elemento.concepto) }}">
+                        class="rup-card {{ elemento.esSolicitud ? 'solicitud' : servicioPrestacion.getCssClass(elemento.concepto) }}">
                         <div class="rup-header">
                             <div class="icon-rup">
-                                <i class="icon icon-rup-semantic-{{prestacionesService.getIcon(elemento.concepto)}}"></i>
+                                <i
+                                    class="icon icon-rup-semantic-{{ servicioPrestacion.getIcon(elemento.concepto) }}"></i>
                             </div>
                             <div class="title">
                                 <ng-container *ngIf="elemento?.nombre">
@@ -40,7 +41,7 @@
                             <div class="actions">
                                 <small *ngIf="elemento?.relacionadoCon && elemento?.relacionadoCon.length > 0"
                                     class="badge badge-info">
-                                    <b> Relacionado con: </b> 
+                                    <b> Relacionado con: </b>
                                     {{ elemento.relacionadoCon[0].concepto.term[0].toUpperCase() + elemento.relacionadoCon[0].concepto.term.slice(1)}}
                                 </small>
                                 <small class="badge badge-info" *ngIf="elemento.esDiagnosticoPrincipal">
@@ -55,12 +56,12 @@
                                 </span>
                             </div>
                             <div class="content">
-                                <div class="row">
+                                <div class="row" *ngIf="paciente">
                                     <div class="col-12">
                                         <rup [elementoRUP]="elementosRUPService.buscarElemento(elemento.concepto, elemento.esSolicitud)"
                                             [params]="elementosRUPService.selectPorRefsetId(elemento.concepto)"
-                                            [prestacion]="prestacion" [registro]="elemento"
-                                            [paciente]="paciente" [soloValores]="true" [vistaHUDS]="false">
+                                            [prestacion]="prestacion" [registro]="elemento" [paciente]="paciente"
+                                            [soloValores]="true" [vistaHUDS]="false">
                                         </rup>
                                     </div>
                                 </div>
