+++ src/app/modules/rup/components/huds/vistaPrestacion.ts
@@ -1,9 +1,11 @@
-import { Component, ViewEncapsulation, Input } from '@angular/core';
-import { IPrestacionRegistro } from '../../interfaces/prestacion.registro.interface';
+import { Component, ViewEncapsulation, Input, HostBinding } from '@angular/core';
 import { IPrestacion } from '../../interfaces/prestacion.interface';
 import { PrestacionesService } from '../../services/prestaciones.service';
 import { ElementosRUPService } from '../../services/elementosRUP.service';
 import { IPaciente } from '../../../../interfaces/IPaciente';
+import { Plex } from '@andes/plex';
+import { Auth } from '@andes/auth';
+import { PacienteService } from '../../../../services/paciente.service';
 
 @Component({
     selector: 'vista-prestacion',
@@ -12,17 +14,42 @@ import { IPaciente } from '../../../../interfaces/IPaciente';
 })
 
 export class VistaPrestacionComponent {
+    @HostBinding('class.plex-layout') layout = true;
 
     @Input() paciente: IPaciente;
     @Input() prestacion: IPrestacion;
     @Input() evolucionActual: any;
     @Input() indice = 0;
 
-    constructor(public prestacionesService: PrestacionesService, public elementosRUPService: ElementosRUPService) {
+    constructor(
+        public servicioPrestacion: PrestacionesService,
+        private servicioPaciente: PacienteService,
+        public elementosRUPService: ElementosRUPService) {
+    }
+
+    private _idPrestacion;
+    @Input()
+    set idPrestacion(value: any) {
+        this.paciente = null;
+        this._idPrestacion = value;
+        this.elementosRUPService.ready.subscribe((resultado) => {
+            if (resultado) {
+                this.servicioPrestacion.getById(this.idPrestacion).subscribe(prestacion => {
+                    this.prestacion = prestacion;
+                    this.servicioPaciente.getById(this.prestacion.paciente.id).subscribe(paciente => {
+                        this.paciente = paciente;
+                    });
+                });
+            }
+        });
+    }
+    get idPrestacion(): any {
+        return this._idPrestacion;
     }
 
     getTimestamp(fecha) {
         return fecha.getTime();
     }
+
 
 }
